Fix "cursors" notification bugs
It seems that the cursors created with the "cursors" library would sometimes
not send position updates properly to their associated onDrag event handler
when moved quickly. Adding an extra call to onDrag in the on-release mouse
event handler fixes this.

Similarly, the "cursors" library seemed to send too many "onReleased" events
--- basically continuously during dragging. Fixed this too.